voice_agent/requirements.txt
anthropic>=0.40.0
mcp>=1.0.0
httpx>=0.27.0
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.14
python-dotenv>=1.0.0
